requirements.txt
PyQt6
requests
openai
python-dotenv
plyer